ML helper app/requirements.txt
streamlit==1.29.0
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1
pandas==2.1.4
streamlit-lottie==0.0.5
streamlit-pills==0.3.0
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
uuid==1.30
json5==0.9.14 